# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mizu water delivery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Order failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{userId}:
    get:
      operationId: getUserOrders
      tags: [orders]
      summary: Get all orders for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (driver)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/accept:
    post:
      operationId: acceptOrder
      tags: [orders]
      summary: Driver accepts a pending order (atomic, prevents collision)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptOrderInput"
      responses:
        "200":
          description: Order accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "409":
          description: Order already taken by another driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/active:
    get:
      operationId: getActiveOrders
      tags: [orders]
      summary: Get all pending orders filtered by driver commune
      parameters:
        - name: driverId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Active orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    OrderInput:
      type: object
      required: [userId, waterVolume, barrelCount, totalPrice]
      properties:
        userId:
          type: string
        waterVolume:
          type: string
        barrelCount:
          type: integer
        totalPrice:
          type: number
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
    AcceptOrderInput:
      type: object
      required: [driverId]
      properties:
        driverId:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: ["معلق", "قيد التوصيل", "وصل السائق", "تم التوصيل"]
    Order:
      type: object
      required: [id, userId, waterVolume, barrelCount, totalPrice, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        driverId:
          type: ["string", "null"]
        userName:
          type: ["string", "null"]
        userPhone:
          type: ["string", "null"]
        waterVolume:
          type: string
        barrelCount:
          type: integer
        totalPrice:
          type: number
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        status:
          type: string
        createdAt:
          type: string
...